Move non render functionality out of module
